Add landing page that displays list of recipes
Create a recipes fragment layout that is to be used as the
landing page. It has a recyclerview to display the list of
recipes. An instance of the recipe view model is binded to it.

Create a recipe list item layout with a Recipe object and a
recipe click listener binded to it.

Create a view model and view model factory to provide access to
the recipe table's LiveData.

Create an adapter that inflates the layout for each recipe list
item and sets the values of the binded Recipe object and click
listener. Instantiates a recipe diff callback that will optimize
updates to the list by calculating the minimum number of
operations needed to update the data.

Within the recipes fragment, get a reference to the binding
object, inflate the fragment views, set the binding object
lifecycle, set the recipes view model, set the recipes adapter.
Then, observe changes to the LiveData of recipes within the view
model; if the LiveData changes, submit the updated list of
recipes.