Consolidate time data into one field, reformat recipe list item
Instead of cook and prep time, have only one total time field
in the recipe table, which is more applicable to a wider array of
recipes.

Reformat the recipe list item to include a timer icon and the
total time, which are right aligned on the same line as the recipe
name.

Remove util functions and strings associated with prep/cook time.