Add night mode colors and update day mode colors
Add night mode colors, which are lighter to go against the darker
background.

Add background and surface colors for day/night mode. Surface
is the color of the list items on the screen.

Update the secondary color to orange.

Add a custom theme for the action bar so that its color can be
grey instead of the primary color.