Add menu item that toggles day/night mode
Save an isDay boolean to a preference so that the day/night
mode setting persists. Add a menu to MainActivity that has
an item that toggles the day/night mode preference when
clicked. Then, a preference listener changes the app from
day/night to night/day.

Add icons for the day/night menu item. Add fade in/out
transitions for activities so that the transition from
day/night to night/day is smooth.

Update the version of androidx.appcompat:appcompat so that
activities are automatically recreated when the day/night
mode changes.