Change adjustPan to adjustResize
This change to windowSoftInputMode is to allow the user to discard
or save without having to back out of the keyboard. This keeps the
discard and save buttons visible at all times; I think the
increased ease of use outweighs the screen space that is taken by
the buttons while the user is typing.

If the keyboard was visible when the user discards/saves, the
keyboard remained open when the app navigated back to the recipe
fragment. So, I use the hideKeyboard funciton to hide the keyboard
if it is showing during navigation.