Trim ingredients/instructions input before filter empties
Previously, when I filtered empty strings before I trimmed, the
user could type a bunch of spaces into a line, which would not
get filtered out, and then it would be trimmed, and then it would
be saved. Thus, a useless string of spaces was saved as "". Now,
when I trim first, useless strings like this are trimmed down to
"" and then filtered out.